# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Franklins AI Driver Onboarding - Development/Demo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers/{id}/tasks:
    get:
      operationId: listDriverTasks
      tags: [tasks]
      summary: List tasks for a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingTask"
    post:
      operationId: createDriverTask
      tags: [tasks]
      summary: Create a task for a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingTask"
  /drivers/{id}/tasks/{taskId}:
    patch:
      operationId: updateDriverTask
      tags: [tasks]
      summary: Update a task for a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: taskId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingTask"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/hired:
    post:
      operationId: simulateHiredEvent
      tags: [events]
      summary: Simulate a Recruiter Hired event (idempotent by externalRecruitId)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HiredEventInput"
      responses:
        "201":
          description: New onboarding record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "200":
          description: Existing record returned (duplicate suppressed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
      required:
        - code
        - message
        - retryable
    Driver:
      type: object
      required:
        - id
        - fullName
        - driverType
        - status
        - stage
        - recruiterName
        - sourceChannel
        - createdAt
        - updatedAt
      properties:
        id:
          type: number
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        driverType:
          type: string
          enum: [owner_operator, company_driver]
        status:
          type: string
          enum: [pre_hire, in_progress, pending_approval, approved, ready_for_dispatch, dispatched, fallout, on_hold]
        stage:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        recruiterName:
          type: string
        sourceChannel:
          type: string
        assigneeId:
          type: ["number", "null"]
        assigneeName:
          type: ["string", "null"]
        operationalOwnerId:
          type: ["number", "null"]
        operationalOwnerName:
          type: ["string", "null"]
        truckVin:
          type: ["string", "null"]
        truckInfo:
          type: ["string", "null"]
        telegramGroupLinked:
          type: boolean
        readyForDispatch:
          type: boolean
        datatruckSyncStatus:
          type: ["string", "null"]
          enum: [pending, synced, failed, retry, "null"]
        startDate:
          type: ["string", "null"]
        slaDeadline:
          type: ["string", "null"]
        blockers:
          type: ["string", "null"]
        nextBestAction:
          type: ["string", "null"]
        externalRecruitId:
          type: ["string", "null"]
        complianceGatesPassed:
          type: boolean
        completionPercent:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
...
    OnboardingTask:
      type: object
      required:
        - id
        - driverId
        - title
        - taskType
        - status
        - createdAt
      properties:
        id:
          type: number
        driverId:
          type: number
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [document_collection, verification, training, admin, communication, compliance]
        status:
          type: string
          enum: [pending, in_progress, blocked, completed, skipped]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assigneeId:
          type: ["number", "null"]
        assigneeName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        isMandatory:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    OnboardingTaskInput:
      type: object
      required:
        - title
        - taskType
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        taskType:
          type: string
          enum: [document_collection, verification, training, admin, communication, compliance]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assigneeId:
          type: number
        dueDate:
          type: string
        isMandatory:
          type: boolean
    OnboardingTaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, blocked, completed, skipped]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assigneeId:
          type: ["number", "null"]
        dueDate:
          type: string
        isMandatory:
          type: boolean
...
    HiredEventInput:
      type: object
      required:
        - fullName
        - driverType
        - recruiterName
        - sourceChannel
        - externalRecruitId
      properties:
        fullName:
          type: string
          minLength: 1
        phone:
          type: string
        email:
          type: string
        state:
          type: string
        driverType:
          type: string
          enum: [owner_operator, company_driver]
        recruiterName:
          type: string
        sourceChannel:
          type: string
        externalRecruitId:
          type: string
          minLength: 1
        truckVin:
          type: string
        assigneeId:
          type: number
        priority:
          type: string
          enum: [critical, high, medium, low]
...